SonarLint suggestion fixes

+++ ESPController/include/influxdb.h
@@ -7,7 +7,7 @@
 
 void influx_task_action();
 
-extern uint16_t TotalNumberOfCells();
+extern uint8_t TotalNumberOfCells();
 extern diybms_eeprom_settings mysettings;
 extern bool wifi_isconnected;
